feat: implement VerseCard components with audio playback and swipe functionality
